+++ package.json
@@ -5,7 +5,6 @@
   "main": "index.ts",
   "scripts": {
     "schedule": "'ts-node' src/index.ts",
-    "activate": "'ts-node' src/authorize.ts",
     "build": "tsc"
   },
   "packageManager": "pnpm@8.6.12",
